+++ src/__tests__/components/FeedItem/VideoStatistics.test.tsx
@@ -1,6 +1,6 @@
 import { shallow } from 'enzyme'
 import VideoStatistics from '../../../components/FeedItem/VideoStatistics'
-import { Feed } from '../../../models/Feed'
+import { Feed } from '../../../domain/Feed'
 
 jest.mock('../../../services/NumberService', () => ({
   formatNumber: (input: number) => input,
