+++ src/__tests__/components/UserTab/UserInfo.test.tsx
@@ -1,6 +1,6 @@
 import Enzyme, { shallow } from 'enzyme'
 import UserInfo from '../../../components/UserTab/UserInfo'
-import { User, UserData, UserStats } from '../../../models/UserData'
+import { User, UserData, UserStats } from '../../../domain/UserData'
 
 jest.mock('../../../services/NumberService', () => ({
   formatNumber: (num: number) => num,
