+++ src/__tests__/hooks/useFeed.test.tsx
@@ -2,7 +2,7 @@ import useFeed from '../../hooks/useFeed'
 import FeedService from '../../services/FeedService'
 import Enzyme, { mount } from 'enzyme'
 import { FC } from 'react'
-import { Feed } from '../../models/Feed'
+import { Feed } from '../../domain/Feed'
 import { act } from 'react-dom/test-utils'
 
 jest.mock('../../services/FeedService')
