+++ src/components/FeedItem/FeedDescription.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react'
 import { Link } from 'react-router-dom'
-import { AuthorMeta } from '../../models/AuthorMeta'
-import { Hashtag } from '../../models/Feed'
+import { AuthorMeta } from '../../domain/AuthorMeta'
+import { Hashtag } from '../../domain/Feed'
 
 interface FeedDescriptionProps {
   authorMeta: AuthorMeta
