+++ src/domain/Feed.ts
